requirements.txt
streamlit
langchain
langchain-google-genai
langchain-community
pandas
sqlalchemy
python-dotenv